feat: change delete student endpoint to GET and add error handling in tests

+++ 11-spring-boot-unit-testing-main/3.00-starting-project/src/main/java/com/luv2code/springmvc/controller/GradebookController.java
@@ -42,10 +42,15 @@ public class GradebookController {
         return "studentInformation";
     }
 
-    @DeleteMapping("delete/student/{id}")
+    @GetMapping("/delete/student/{id}")
     public String deleteStudent(@PathVariable int id) {
+
+        if (!studentAndGradeService.checkIfStudentIsNull(id)) {
+            return "error";
+        }
+
         studentAndGradeService.deleteStudent(id);
-        return "index";
+        return "redirect:/";
     }
 
 }
